testing added double

+++ Assignment #4 paging/src/Paging.java
@@ -13,9 +13,9 @@ public class Paging {
 		
 		//returns the hit ratio
 		double h1 = fifo.run(); 
-		int h2 = lfu.run(); 
-		int h3 = lru.run(); 
-		int h4 = mfu.run(); 
+		double h2 = lfu.run(); 
+		double h3 = lru.run(); 
+		double h4 = mfu.run(); 
 		
 		//print average hit ratio
 		Print.thisStringln("--Average hit ratios--");
